Changed boards/pedals length/width attr to float
For whatever reason, decimal type was not working, possibly because
it isn't a type in JavaScript (really??) so I have changed all
length and width attributes for both boards and pedals to float,
and now the curl requests work just fine and save as decimals!

+++ db/migrate/20170113162921_create_boards.rb
@@ -2,8 +2,8 @@ class CreateBoards < ActiveRecord::Migration
   def change
     create_table :boards do |t|
       t.string :name
-      t.decimal :length
-      t.decimal :width
+      t.float :length
+      t.float :width
 
       t.timestamps null: false
     end
